Add SPHINCS-256, extracted from SUPERCOP-20171218
As the SUPERCOP directories do not always contain specific
implementations of chacha12, blake256 and blake512, we include the
respective reference implementations as present in SUPERCOP.

We also omit the dependency on openssl; it was unused, but
complicates compilation on systems that do not have it available.

The test/ files were provided by the SPHINCS-256 implementors.